Make Rails do it automagically.
If a REVISION file is found in the project root no initializer is
required, this makes it easy to just include in your Gemfile and have
the basics up and running.

If you do specify a custom content, the sha will still be read from the
file, but if you do decided to override it the sha won't be read.